cli: port merge onto the shared rewrite engine
Move the merge pipeline into the rewrite module and share the low-level
machinery (writer setup, summary/index inspection, checked slicing, and the
metadata/attachment index-or-scan traversals) between the single-input rewrite
commands and merge via a new rewrite::common module.

- Extract rewrite/common.rs with helpers previously duplicated across the merge
  command and the rewrite engine.
- Move the k-way merge (heap, cross-input schema/channel remapping, coalescing,
  metadata dedup) into rewrite/merge.rs; commands/merge.rs is now a thin adapter.
- Fix a latent data-loss bug: the single-input indexed path now falls back to a
  lossless linear scan when the summary does not index every message (loose
  top-level messages or missing message indexes), matching merge and preventing
  filter/sort/compress/decompress from silently dropping such records.
- Unify merge's output flag with the other rewrite commands: --output (-o) is
  now canonical and --output-file is a hidden, warned deprecated alias.

Record placement, message ordering, coalescing, metadata/attachment handling,
and the in-memory buffering profile are unchanged; bounded-memory ordering for
summaryless inputs remains a follow-up.
