+++ rust/cli/src/commands/merge.rs
@@ -1,202 +1,27 @@
-use std::cmp::Ordering;
-use std::collections::{BTreeMap, BinaryHeap, HashMap, HashSet};
-use std::io::{IsTerminal as _, Seek, Write};
-use std::path::PathBuf;
-use std::sync::Arc;
+//! The `merge` command: k-way merge several MCAP inputs into one output ordered by log time. This
+//! is a thin adapter over the shared [`crate::rewrite`] merge pipeline, which owns the read/merge/
+//! write machinery and the standardized record placement.
+use anyhow::Result;
 
-use anyhow::{bail, Context, Result};
-use mcap::records::{MessageHeader, Record};
-
-use crate::cli::{CoalesceChannels, MergeCommand};
+use crate::cli::{warn_output_file_deprecation, MergeCommand};
 use crate::context::CommandContext;
-
-#[derive(Debug, Clone)]
-struct MergeOptions {
-    files: Vec<PathBuf>,
-    output_file: Option<PathBuf>,
-    compression: Option<mcap::Compression>,
-    chunk_size: u64,
-    include_crc: bool,
-    chunked: bool,
-    allow_duplicate_metadata: bool,
-    coalesce_channels: CoalesceChannels,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct SchemaKey {
-    name: String,
-    encoding: String,
-    data: Vec<u8>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct ChannelKey {
-    schema_id: u16,
-    topic: String,
-    message_encoding: String,
-    metadata: Vec<(String, String)>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-struct MetadataKey {
-    name: String,
-    entries: Vec<(String, String)>,
-}
-
-#[derive(Debug, Clone)]
-struct InputRef<'a> {
-    name: &'a str,
-    data: &'a [u8],
-}
-
-struct InputMessageReader<'a> {
-    name: &'a str,
-    records: mcap::read::ChunkFlattener<'a>,
-    schemas: HashMap<u16, Arc<mcap::Schema<'static>>>,
-    channels: HashMap<u16, Arc<mcap::Channel<'static>>>,
-}
-
-struct IndexedInputMessageReader<'a> {
-    input_idx: usize,
-    input_order: usize,
-    name: &'a str,
-    data: &'a [u8],
-    summary: Box<mcap::Summary>,
-    reader: mcap::sans_io::IndexedReader,
-}
-
-struct MaterializedInputMessages {
-    messages: Vec<PendingMessage>,
-    next: usize,
-}
-
-enum MergeMessageStream<'a> {
-    Indexed(IndexedInputMessageReader<'a>),
-    Materialized(MaterializedInputMessages),
-}
-
-type InputMessage = (Arc<mcap::Channel<'static>>, MessageHeader, Vec<u8>);
-const MESSAGE_INDEX_ENTRY_SIZE: usize = 16;
-
-#[derive(Debug, Clone)]
-struct PendingMessage {
-    input_idx: usize,
-    input_order: usize,
-    input_channel_id: u16,
-    channel: Arc<mcap::Channel<'static>>,
-    sequence: u32,
-    log_time: u64,
-    publish_time: u64,
-    data: Vec<u8>,
-}
-
-impl PendingMessage {
-    fn new(
-        input_idx: usize,
-        input_order: usize,
-        channel: Arc<mcap::Channel<'static>>,
-        header: MessageHeader,
-        data: Vec<u8>,
-    ) -> Self {
-        Self {
-            input_idx,
-            input_order,
-            input_channel_id: header.channel_id,
-            channel,
-            sequence: header.sequence,
-            log_time: header.log_time,
-            publish_time: header.publish_time,
-            data,
-        }
-    }
-}
-
-impl PartialEq for PendingMessage {
-    fn eq(&self, other: &Self) -> bool {
-        self.log_time == other.log_time
-            && self.input_idx == other.input_idx
-            && self.input_order == other.input_order
-    }
-}
-
-impl Eq for PendingMessage {}
-
-impl PartialOrd for PendingMessage {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for PendingMessage {
-    fn cmp(&self, other: &Self) -> Ordering {
-        other
-            .log_time
-            .cmp(&self.log_time)
-            .then_with(|| other.input_idx.cmp(&self.input_idx))
-            .then_with(|| other.input_order.cmp(&self.input_order))
-    }
-}
-
-#[derive(Default)]
-struct MetadataState {
-    seen_metadata: HashSet<MetadataKey>,
-    metadata_names: HashSet<String>,
-}
-
-#[derive(Default)]
-struct IdMaps {
-    schema_ids: HashMap<(usize, u16), u16>,
-    schema_ids_by_content: HashMap<SchemaKey, u16>,
-    channel_ids: HashMap<(usize, u16), u16>,
-    channel_ids_by_content: HashMap<ChannelKey, u16>,
-    next_output_channel_id: u16,
-}
+use crate::rewrite::{self, MergeOptions};
+use crate::source::SourceOptions;
 
 pub fn run(ctx: &CommandContext, args: MergeCommand) -> Result<()> {
-    let opts = build_merge_options(args);
-    let source_options = crate::source::SourceOptions::new(ctx.allow_remote_scan());
-
-    if let Some(output_path) = &opts.output_file {
-        for input_path in &opts.files {
-            crate::source::ensure_distinct_local_input_output(input_path, output_path)?;
-        }
-    }
-
-    let mut mapped_inputs = Vec::with_capacity(opts.files.len());
-    let mut input_names = Vec::with_capacity(opts.files.len());
-    for path in &opts.files {
-        let mapped = crate::source::load_path(path, source_options)?;
-        mapped_inputs.push(mapped);
-        input_names.push(crate::source::redacted_display(path));
-    }
-
-    let input_refs: Vec<InputRef<'_>> = mapped_inputs
-        .iter()
-        .zip(input_names.iter())
-        .map(|(mapped, name)| InputRef {
-            name: name.as_str(),
-            data: mapped.as_slice(),
-        })
-        .collect();
-
-    if let Some(output_path) = &opts.output_file {
-        let output = std::fs::File::create(output_path)
-            .with_context(|| format!("failed to open '{}' for writing", output_path.display()))?;
-        merge_inputs(&input_refs, output, &opts, false)
-    } else {
-        if std::io::stdout().is_terminal() {
-            bail!("{}", crate::source::PLEASE_REDIRECT);
-        }
-        let stdout = std::io::stdout();
-        let output = mcap::write::NoSeek::new(stdout.lock());
-        merge_inputs(&input_refs, output, &opts, true)
-    }
+    warn_output_file_deprecation(args.output_file.as_deref());
+    rewrite::run_merge(
+        build_merge_options(args),
+        SourceOptions::new(ctx.allow_remote_scan()),
+    )
 }
 
+/// Maps the parsed `merge` CLI arguments onto the engine-facing [`MergeOptions`]. The output path
+/// prefers `--output`, falling back to the deprecated `--output-file` alias.
 fn build_merge_options(args: MergeCommand) -> MergeOptions {
     MergeOptions {
         files: args.files,
-        output_file: args.output_file,
+        output: args.output.or(args.output_file),
         compression: args.compression.to_compression(),
         chunk_size: args.chunk_size,
         include_crc: !args.no_crc,
@@ -206,930 +31,19 @@ fn build_merge_options(args: MergeCommand) -> MergeOptions {
     }
 }
 
-fn merge_inputs<W: Write + Seek>(
-    inputs: &[InputRef<'_>],
-    sink: W,
-    opts: &MergeOptions,
-    disable_seeking: bool,
-) -> Result<()> {
-    let profiles = inputs
-        .iter()
-        .map(read_profile)
-        .collect::<Result<Vec<_>>>()?;
-    let output_profile = output_profile(&profiles);
-
-    let mut write_options = mcap::WriteOptions::new()
-        .profile(output_profile)
-        .library(crate::cli::LIBRARY_IDENTIFIER.clone())
-        .use_chunks(opts.chunked)
-        .chunk_size(Some(opts.chunk_size))
-        .compression(opts.compression)
-        .calculate_chunk_crcs(opts.include_crc)
-        .calculate_data_section_crc(opts.include_crc)
-        .calculate_summary_section_crc(opts.include_crc)
-        .calculate_attachment_crcs(opts.include_crc)
-        .disable_seeking(disable_seeking);
-
-    if !opts.chunked {
-        write_options = write_options.emit_message_indexes(false);
-    }
-
-    let mut writer = write_options
-        .create(sink)
-        .context("failed to create mcap writer")?;
-
-    let summaries = inputs
-        .iter()
-        // Treat summary lookup as best effort and fall back to linear scans when
-        // summary parsing fails.
-        .map(|input| mcap::Summary::read(input.data).unwrap_or_default())
-        .collect::<Vec<_>>();
-
-    merge_messages(
-        inputs,
-        &summaries,
-        &mut writer,
-        opts.coalesce_channels,
-        opts.allow_duplicate_metadata,
-    )?;
-
-    for (idx, input) in inputs.iter().enumerate() {
-        write_attachments(&mut writer, input, summaries[idx].as_ref())?;
-    }
-
-    writer.finish().context("failed to finish mcap writer")?;
-    Ok(())
-}
-
-fn read_profile(input: &InputRef<'_>) -> Result<String> {
-    let mut reader = mcap::read::LinearReader::new(input.data)
-        .with_context(|| format!("failed to read '{}'", input.name))?;
-    match reader.next() {
-        Some(Ok(Record::Header(header))) => Ok(header.profile),
-        Some(Ok(_)) | None => Ok(String::new()),
-        Some(Err(err)) => Err(anyhow::Error::from(err))
-            .with_context(|| format!("failed to read header from '{}'", input.name)),
-    }
-}
-
-fn output_profile(profiles: &[String]) -> String {
-    let Some(first) = profiles.first() else {
-        return String::new();
-    };
-    if profiles.iter().all(|profile| profile == first) {
-        first.clone()
-    } else {
-        String::new()
-    }
-}
-
-fn write_merged_metadata<W: Write + Seek>(
-    writer: &mut mcap::Writer<W>,
-    state: &mut MetadataState,
-    metadata_record: mcap::records::Metadata,
-    allow_duplicate_metadata: bool,
-) -> Result<()> {
-    if state.metadata_names.contains(&metadata_record.name) && !allow_duplicate_metadata {
-        bail!(
-            "metadata name '{}' was previously encountered. Supply --allow-duplicate-metadata to override.",
-            metadata_record.name
-        );
-    }
-
-    let key = metadata_key(&metadata_record);
-    if state.seen_metadata.insert(key) {
-        writer.write_metadata(&metadata_record)?;
-        state.metadata_names.insert(metadata_record.name.clone());
-    }
-    Ok(())
-}
-
-fn metadata_key(metadata_record: &mcap::records::Metadata) -> MetadataKey {
-    MetadataKey {
-        name: metadata_record.name.clone(),
-        entries: metadata_record
-            .metadata
-            .iter()
-            .map(|(key, value)| (key.clone(), value.clone()))
-            .collect(),
-    }
-}
-
-fn write_attachments<W: Write + Seek>(
-    writer: &mut mcap::Writer<W>,
-    input: &InputRef<'_>,
-    summary: Option<&mcap::Summary>,
-) -> Result<()> {
-    if let Some(summary) = summary {
-        let attachment_count = summary.stats.as_ref().map(|stats| stats.attachment_count);
-        if let Some(attachment_count) = attachment_count {
-            if attachment_count == summary.attachment_indexes.len() as u32 {
-                let mut indexes = summary.attachment_indexes.clone();
-                indexes.sort_by_key(|index| index.offset);
-                for index in indexes {
-                    let attachment =
-                        mcap::read::attachment(input.data, &index).with_context(|| {
-                            format!(
-                                "failed to read attachment '{}' at offset {} from '{}'",
-                                index.name, index.offset, input.name
-                            )
-                        })?;
-                    writer.attach(&attachment).with_context(|| {
-                        format!(
-                            "failed to write attachment '{}' from '{}'",
-                            index.name, input.name
-                        )
-                    })?;
-                }
-                return Ok(());
-            }
-        }
-    }
-
-    for record in mcap::read::LinearReader::new(input.data)
-        .with_context(|| format!("failed to read '{}'", input.name))?
-    {
-        if let Record::Attachment { header, data, .. } =
-            record.with_context(|| format!("failed to parse '{}'", input.name))?
-        {
-            writer.attach(&mcap::Attachment {
-                log_time: header.log_time,
-                create_time: header.create_time,
-                name: header.name,
-                media_type: header.media_type,
-                data: std::borrow::Cow::Borrowed(data.as_ref()),
-            })?;
-        }
-    }
-
-    Ok(())
-}
-
-fn write_metadata_records<W: Write + Seek>(
-    writer: &mut mcap::Writer<W>,
-    state: &mut MetadataState,
-    input: &InputRef<'_>,
-    summary: Option<&mcap::Summary>,
-    allow_duplicate_metadata: bool,
-) -> Result<()> {
-    if let Some(summary) = summary {
-        let metadata_count = summary.stats.as_ref().map(|stats| stats.metadata_count);
-        if metadata_count == Some(summary.metadata_indexes.len() as u32) {
-            let mut indexes = summary.metadata_indexes.clone();
-            indexes.sort_by_key(|index| index.offset);
-            for index in indexes {
-                let metadata = mcap::read::metadata(input.data, &index).with_context(|| {
-                    format!(
-                        "failed to read metadata '{}' at offset {} from '{}'",
-                        index.name, index.offset, input.name
-                    )
-                })?;
-                write_merged_metadata(writer, state, metadata, allow_duplicate_metadata)?;
-            }
-            return Ok(());
-        }
-    }
-
-    for record in mcap::read::LinearReader::new(input.data)
-        .with_context(|| format!("failed to read '{}'", input.name))?
-    {
-        if let Record::Metadata(metadata) =
-            record.with_context(|| format!("failed to parse '{}'", input.name))?
-        {
-            write_merged_metadata(writer, state, metadata, allow_duplicate_metadata)?;
-        }
-    }
-
-    Ok(())
-}
-
-fn summary_indexes_all_messages(input: &[u8], summary: &mcap::Summary) -> bool {
-    let Some(stats) = summary.stats.as_ref() else {
-        return false;
-    };
-    let mut offsets = HashSet::new();
-    let mut indexed_messages = 0u64;
-    for offset in summary
-        .chunk_indexes
-        .iter()
-        .flat_map(|chunk| chunk.message_index_offsets.values())
-    {
-        if offsets.insert(*offset) {
-            let Ok(count) = message_index_count(input, *offset) else {
-                return false;
-            };
-            indexed_messages = indexed_messages.saturating_add(count as u64);
-        }
-    }
-    indexed_messages == stats.message_count
-}
-
-fn message_index_count(input: &[u8], offset: u64) -> Result<usize> {
-    let start = usize::try_from(offset).with_context(|| {
-        format!("message index offset out of range for this platform: {offset}")
-    })?;
-    let header_end = start
-        .checked_add(9)
-        .ok_or_else(|| anyhow::anyhow!("message index header overflows at offset {offset}"))?;
-    let header = input
-        .get(start..header_end)
-        .ok_or_else(|| anyhow::anyhow!("message index header out of bounds at offset {offset}"))?;
-    let opcode = header[0];
-    if opcode != mcap::records::op::MESSAGE_INDEX {
-        bail!("expected MessageIndex record at offset {offset}");
-    }
-    let length = u64::from_le_bytes(header[1..9].try_into().expect("slice has length 8"));
-    let length = usize::try_from(length)
-        .with_context(|| format!("message index length out of range at offset {offset}"))?;
-    let body_start = header_end;
-    let body_end = body_start
-        .checked_add(length)
-        .ok_or_else(|| anyhow::anyhow!("message index length overflows at offset {offset}"))?;
-    let body = input
-        .get(body_start..body_end)
-        .ok_or_else(|| anyhow::anyhow!("message index body out of bounds at offset {offset}"))?;
-
-    if body.len() < 6 {
-        bail!("message index body too short at offset {offset}");
-    }
-    let byte_len = u32::from_le_bytes(body[2..6].try_into().expect("slice has length 4")) as usize;
-    if !byte_len.is_multiple_of(MESSAGE_INDEX_ENTRY_SIZE) {
-        bail!("message index entries are misaligned at offset {offset}");
-    }
-    let expected_len = 6usize
-        .checked_add(byte_len)
-        .ok_or_else(|| anyhow::anyhow!("message index byte length overflows at offset {offset}"))?;
-    if body.len() != expected_len {
-        bail!("message index length mismatch at offset {offset}");
-    }
-    Ok(byte_len / MESSAGE_INDEX_ENTRY_SIZE)
-}
-
-impl<'a> InputMessageReader<'a> {
-    fn new(input: &InputRef<'a>) -> Result<Self> {
-        Ok(Self {
-            name: input.name,
-            records: mcap::read::ChunkFlattener::new(input.data)
-                .with_context(|| format!("failed to stream records from '{}'", input.name))?,
-            schemas: HashMap::new(),
-            channels: HashMap::new(),
-        })
-    }
-}
-
-fn next_message_from_input(input: &mut InputMessageReader<'_>) -> Result<Option<InputMessage>> {
-    for record in input.records.by_ref() {
-        let record = record.with_context(|| format!("failed to parse '{}'", input.name))?;
-        match record {
-            Record::Schema { header, data } => {
-                let schema = Arc::new(mcap::Schema {
-                    id: header.id,
-                    name: header.name.clone(),
-                    encoding: header.encoding.clone(),
-                    data: std::borrow::Cow::Owned(data.into_owned()),
-                });
-                if let Some(existing) = input.schemas.get(&header.id) {
-                    if existing.name != schema.name
-                        || existing.encoding != schema.encoding
-                        || existing.data.as_ref() != schema.data.as_ref()
-                    {
-                        return Err(mcap::McapError::ConflictingSchemas(header.name).into());
-                    }
-                } else {
-                    input.schemas.insert(header.id, schema);
-                }
-            }
-            Record::Channel(channel) => {
-                let schema = if channel.schema_id == 0 {
-                    None
-                } else {
-                    Some(
-                        input
-                            .schemas
-                            .get(&channel.schema_id)
-                            .cloned()
-                            .ok_or_else(|| {
-                                anyhow::anyhow!(
-                                    "encountered channel '{}' with unknown schema {} in '{}'",
-                                    channel.topic,
-                                    channel.schema_id,
-                                    input.name
-                                )
-                            })?,
-                    )
-                };
-                let parsed_channel = Arc::new(mcap::Channel {
-                    id: channel.id,
-                    topic: channel.topic.clone(),
-                    schema: schema.clone(),
-                    message_encoding: channel.message_encoding.clone(),
-                    metadata: channel.metadata.clone(),
-                });
-
-                if let Some(existing) = input.channels.get(&channel.id) {
-                    if existing.topic != parsed_channel.topic
-                        || existing.schema.as_ref().map(|schema| schema.id)
-                            != parsed_channel.schema.as_ref().map(|schema| schema.id)
-                        || existing.message_encoding != parsed_channel.message_encoding
-                        || existing.metadata != parsed_channel.metadata
-                    {
-                        return Err(mcap::McapError::ConflictingChannels(channel.topic).into());
-                    }
-                } else {
-                    input.channels.insert(channel.id, parsed_channel);
-                }
-            }
-            Record::Message { header, data } => {
-                let channel = input
-                    .channels
-                    .get(&header.channel_id)
-                    .cloned()
-                    .ok_or_else(|| {
-                        anyhow::anyhow!(
-                            "encountered message referencing unknown channel {} in '{}'",
-                            header.channel_id,
-                            input.name
-                        )
-                    })?;
-                return Ok(Some((channel, header, data.into_owned())));
-            }
-            _ => {}
-        }
-    }
-    Ok(None)
-}
-
-impl<'a> IndexedInputMessageReader<'a> {
-    fn new(input_idx: usize, input: &InputRef<'a>, summary: mcap::Summary) -> Result<Self> {
-        let reader = mcap::sans_io::IndexedReader::new_with_options(
-            &summary,
-            mcap::sans_io::IndexedReaderOptions::new()
-                .with_order(mcap::sans_io::indexed_reader::ReadOrder::LogTime),
-        )
-        .with_context(|| format!("failed to create indexed reader for '{}'", input.name))?;
-
-        Ok(Self {
-            input_idx,
-            input_order: 0,
-            name: input.name,
-            data: input.data,
-            summary: Box::new(summary),
-            reader,
-        })
-    }
-
-    fn next_message(&mut self) -> Result<Option<PendingMessage>> {
-        while let Some(event) = self.reader.next_event() {
-            match event.with_context(|| format!("failed to read indexed '{}'", self.name))? {
-                mcap::sans_io::IndexedReadEvent::ReadChunkRequest { offset, length } => {
-                    let chunk_data = checked_slice(self.data, offset, length)?;
-                    self.reader.insert_chunk_record_data(offset, chunk_data)?;
-                }
-                mcap::sans_io::IndexedReadEvent::Message { header, data } => {
-                    let channel = self
-                        .summary
-                        .channels
-                        .get(&header.channel_id)
-                        .cloned()
-                        .ok_or_else(|| {
-                            anyhow::anyhow!(
-                                "message references unknown channel {} in '{}'",
-                                header.channel_id,
-                                self.name
-                            )
-                        })?;
-                    let input_order = self.input_order;
-                    self.input_order += 1;
-                    return Ok(Some(PendingMessage::new(
-                        self.input_idx,
-                        input_order,
-                        channel,
-                        header,
-                        data.to_vec(),
-                    )));
-                }
-            }
-        }
-
-        Ok(None)
-    }
-}
-
-impl MaterializedInputMessages {
-    fn new(input_idx: usize, input: &InputRef<'_>) -> Result<Self> {
-        let mut reader = InputMessageReader::new(input)?;
-        let mut messages = Vec::new();
-        let mut input_order = 0usize;
-        while let Some((channel, header, data)) = next_message_from_input(&mut reader)
-            .with_context(|| format!("failed reading messages from '{}'", input.name))?
-        {
-            messages.push(PendingMessage::new(
-                input_idx,
-                input_order,
-                channel,
-                header,
-                data,
-            ));
-            input_order += 1;
-        }
-        messages.sort_by_key(|message| (message.log_time, message.input_order));
-
-        Ok(Self { messages, next: 0 })
-    }
-
-    fn next_message(&mut self) -> Option<PendingMessage> {
-        let message = self.messages.get(self.next).cloned()?;
-        self.next += 1;
-        Some(message)
-    }
-}
-
-impl MergeMessageStream<'_> {
-    fn next_message(&mut self) -> Result<Option<PendingMessage>> {
-        match self {
-            MergeMessageStream::Indexed(reader) => reader.next_message(),
-            MergeMessageStream::Materialized(messages) => Ok(messages.next_message()),
-        }
-    }
-}
-
-fn checked_slice(input: &[u8], offset: u64, length: usize) -> Result<&[u8]> {
-    let start = usize::try_from(offset)
-        .with_context(|| format!("chunk offset out of range for this platform: {offset}"))?;
-    let end = start
-        .checked_add(length)
-        .ok_or_else(|| anyhow::anyhow!("chunk read overflow at offset {offset}"))?;
-    input.get(start..end).ok_or_else(|| {
-        anyhow::anyhow!("chunk read out of bounds at offset {offset} length {length}")
-    })
-}
-
-fn merge_messages<W: Write + Seek>(
-    inputs: &[InputRef<'_>],
-    summaries: &[Option<mcap::Summary>],
-    writer: &mut mcap::Writer<W>,
-    coalesce_channels: CoalesceChannels,
-    allow_duplicate_metadata: bool,
-) -> Result<()> {
-    let mut id_maps = IdMaps {
-        next_output_channel_id: 1,
-        ..IdMaps::default()
-    };
-    let mut metadata_state = MetadataState::default();
-
-    for (input_idx, input) in inputs.iter().enumerate() {
-        write_metadata_records(
-            writer,
-            &mut metadata_state,
-            input,
-            summaries[input_idx].as_ref(),
-            allow_duplicate_metadata,
-        )?;
-    }
-
-    let mut streams = Vec::<MergeMessageStream<'_>>::with_capacity(inputs.len());
-    for (input_idx, input) in inputs.iter().enumerate() {
-        if let Some(summary) = summaries[input_idx].as_ref() {
-            if !summary.chunk_indexes.is_empty()
-                && crate::rewrite::summary_supports_indexed_read(summary)
-                && summary_indexes_all_messages(input.data, summary)
-            {
-                streams.push(MergeMessageStream::Indexed(IndexedInputMessageReader::new(
-                    input_idx,
-                    input,
-                    summary.clone(),
-                )?));
-                continue;
-            }
-        }
-        // Without usable message indexes, guaranteeing log-time order requires sorting this input.
-        streams.push(MergeMessageStream::Materialized(
-            MaterializedInputMessages::new(input_idx, input)?,
-        ));
-    }
-
-    let mut heap = BinaryHeap::<PendingMessage>::new();
-    for stream in &mut streams {
-        if let Some(message) = stream.next_message()? {
-            heap.push(message);
-        }
-    }
-
-    while let Some(message) = heap.pop() {
-        let input_idx = message.input_idx;
-        let output_channel_id = ensure_output_channel_id(
-            &mut id_maps,
-            writer,
-            input_idx,
-            message.input_channel_id,
-            &message.channel,
-            coalesce_channels,
-        )?;
-
-        writer.write_to_known_channel(
-            &MessageHeader {
-                channel_id: output_channel_id,
-                sequence: message.sequence,
-                log_time: message.log_time,
-                publish_time: message.publish_time,
-            },
-            &message.data,
-        )?;
-
-        if let Some(next) = streams[input_idx].next_message()? {
-            heap.push(next);
-        }
-    }
-
-    Ok(())
-}
-
-fn ensure_output_channel_id<W: Write + Seek>(
-    id_maps: &mut IdMaps,
-    writer: &mut mcap::Writer<W>,
-    input_idx: usize,
-    input_channel_id: u16,
-    channel: &Arc<mcap::Channel<'_>>,
-    coalesce_channels: CoalesceChannels,
-) -> Result<u16> {
-    if let Some(output_channel_id) = id_maps.channel_ids.get(&(input_idx, input_channel_id)) {
-        return Ok(*output_channel_id);
-    }
-
-    let output_schema_id = if let Some(schema) = channel.schema.as_ref() {
-        ensure_output_schema_id(id_maps, writer, input_idx, schema)?
-    } else {
-        0
-    };
-
-    if coalesce_channels != CoalesceChannels::None {
-        let channel_key = make_channel_key(
-            output_schema_id,
-            &channel.topic,
-            &channel.message_encoding,
-            &channel.metadata,
-            coalesce_channels,
-        );
-        if let Some(output_channel_id) = id_maps.channel_ids_by_content.get(&channel_key).copied() {
-            id_maps
-                .channel_ids
-                .insert((input_idx, input_channel_id), output_channel_id);
-            return Ok(output_channel_id);
-        }
-
-        let output_channel_id = reserve_next_channel_id(id_maps)?;
-        writer.add_channel_with_id(
-            output_channel_id,
-            output_schema_id,
-            &channel.topic,
-            &channel.message_encoding,
-            &channel.metadata,
-        )?;
-        id_maps
-            .channel_ids
-            .insert((input_idx, input_channel_id), output_channel_id);
-        id_maps
-            .channel_ids_by_content
-            .insert(channel_key, output_channel_id);
-        return Ok(output_channel_id);
-    }
-
-    let output_channel_id = reserve_next_channel_id(id_maps)?;
-    writer.add_channel_with_id(
-        output_channel_id,
-        output_schema_id,
-        &channel.topic,
-        &channel.message_encoding,
-        &channel.metadata,
-    )?;
-    id_maps
-        .channel_ids
-        .insert((input_idx, input_channel_id), output_channel_id);
-    Ok(output_channel_id)
-}
-
-fn ensure_output_schema_id<W: Write + Seek>(
-    id_maps: &mut IdMaps,
-    writer: &mut mcap::Writer<W>,
-    input_idx: usize,
-    schema: &Arc<mcap::Schema<'_>>,
-) -> Result<u16> {
-    if let Some(output_schema_id) = id_maps.schema_ids.get(&(input_idx, schema.id)) {
-        return Ok(*output_schema_id);
-    }
-
-    let key = SchemaKey {
-        name: schema.name.clone(),
-        encoding: schema.encoding.clone(),
-        data: schema.data.clone().into_owned(),
-    };
-
-    let output_schema_id = if let Some(existing_schema_id) = id_maps.schema_ids_by_content.get(&key)
-    {
-        *existing_schema_id
-    } else {
-        let id = writer.add_schema(&schema.name, &schema.encoding, schema.data.as_ref())?;
-        id_maps.schema_ids_by_content.insert(key, id);
-        id
-    };
-
-    id_maps
-        .schema_ids
-        .insert((input_idx, schema.id), output_schema_id);
-    Ok(output_schema_id)
-}
-
-fn reserve_next_channel_id(id_maps: &mut IdMaps) -> Result<u16> {
-    if id_maps.next_output_channel_id == 0 {
-        bail!("cannot merge more than 65535 channels");
-    }
-    let id = id_maps.next_output_channel_id;
-    id_maps.next_output_channel_id = id_maps.next_output_channel_id.wrapping_add(1);
-    Ok(id)
-}
-
-fn make_channel_key(
-    schema_id: u16,
-    topic: &str,
-    message_encoding: &str,
-    metadata: &BTreeMap<String, String>,
-    coalesce_channels: CoalesceChannels,
-) -> ChannelKey {
-    let metadata_entries = if coalesce_channels == CoalesceChannels::Force {
-        Vec::new()
-    } else {
-        metadata
-            .iter()
-            .map(|(key, value)| (key.clone(), value.clone()))
-            .collect()
-    };
-
-    ChannelKey {
-        schema_id,
-        topic: topic.to_string(),
-        message_encoding: message_encoding.to_string(),
-        metadata: metadata_entries,
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeMap;
-    use std::io::Cursor;
+    use std::path::PathBuf;
 
-    use super::*;
-    use crate::cli::CompressionFormat;
-
-    #[derive(Debug, Clone)]
-    struct TestMessage {
-        channel_id: u16,
-        topic: String,
-        metadata: BTreeMap<String, String>,
-        log_time: u64,
-        payload: Vec<u8>,
-    }
-
-    fn build_mcap(
-        profile: &str,
-        messages: &[TestMessage],
-        metadata_records: &[mcap::records::Metadata],
-        attachments: &[mcap::Attachment<'_>],
-        emit_attachment_indexes: bool,
-        emit_metadata_indexes: bool,
-    ) -> Vec<u8> {
-        build_mcap_with_options(
-            profile,
-            messages,
-            metadata_records,
-            attachments,
-            emit_attachment_indexes,
-            emit_metadata_indexes,
-            true,
-        )
-    }
-
-    fn build_mcap_with_options(
-        profile: &str,
-        messages: &[TestMessage],
-        metadata_records: &[mcap::records::Metadata],
-        attachments: &[mcap::Attachment<'_>],
-        emit_attachment_indexes: bool,
-        emit_metadata_indexes: bool,
-        emit_statistics: bool,
-    ) -> Vec<u8> {
-        let mut output = Cursor::new(Vec::<u8>::new());
-        {
-            let mut writer = mcap::WriteOptions::new()
-                .profile(profile)
-                .library("test-recorder/0.0")
-                .emit_attachment_indexes(emit_attachment_indexes)
-                .emit_metadata_indexes(emit_metadata_indexes)
-                .emit_statistics(emit_statistics)
-                .create(&mut output)
-                .expect("writer");
-
-            let mut channels = BTreeMap::<u16, std::sync::Arc<mcap::Channel<'static>>>::new();
-            let schema = std::sync::Arc::new(mcap::Schema {
-                id: 1,
-                name: "example".to_string(),
-                encoding: "jsonschema".to_string(),
-                data: std::borrow::Cow::Borrowed(br#"{"type":"object"}"#),
-            });
-
-            for message in messages {
-                let channel = channels
-                    .entry(message.channel_id)
-                    .or_insert_with(|| {
-                        std::sync::Arc::new(mcap::Channel {
-                            id: message.channel_id,
-                            topic: message.topic.clone(),
-                            schema: Some(schema.clone()),
-                            message_encoding: "json".to_string(),
-                            metadata: message.metadata.clone(),
-                        })
-                    })
-                    .clone();
-                writer
-                    .write(&mcap::Message {
-                        channel,
-                        sequence: message.log_time as u32,
-                        log_time: message.log_time,
-                        publish_time: message.log_time,
-                        data: std::borrow::Cow::Borrowed(message.payload.as_slice()),
-                    })
-                    .expect("write message");
-            }
-
-            for metadata in metadata_records {
-                writer.write_metadata(metadata).expect("write metadata");
-            }
-            for attachment in attachments {
-                writer.attach(attachment).expect("write attachment");
-            }
-
-            writer.finish().expect("finish");
-        }
-        output.into_inner()
-    }
-
-    fn build_non_indexed_mcap(profile: &str, messages: &[TestMessage]) -> Vec<u8> {
-        let mut output = Cursor::new(Vec::<u8>::new());
-        {
-            let mut writer = mcap::WriteOptions::new()
-                .profile(profile)
-                .library("test-recorder/0.0")
-                .use_chunks(false)
-                .emit_summary_records(false)
-                .emit_summary_offsets(false)
-                .create(&mut output)
-                .expect("writer");
-            let schema_id = writer
-                .add_schema("example", "jsonschema", br#"{"type":"object"}"#)
-                .expect("schema");
-            let mut channels = BTreeMap::new();
-
-            for message in messages {
-                let channel_id = *channels.entry(message.channel_id).or_insert_with(|| {
-                    writer
-                        .add_channel(schema_id, &message.topic, "json", &message.metadata)
-                        .expect("channel")
-                });
-                writer
-                    .write_to_known_channel(
-                        &MessageHeader {
-                            channel_id,
-                            sequence: message.log_time as u32,
-                            log_time: message.log_time,
-                            publish_time: message.log_time,
-                        },
-                        message.payload.as_slice(),
-                    )
-                    .expect("write message");
-            }
-
-            writer.finish().expect("finish");
-        }
-        output.into_inner()
-    }
-
-    fn build_indexed_mcap_with_loose_message() -> Vec<u8> {
-        let mut output = Cursor::new(Vec::<u8>::new());
-        let channel_id;
-        {
-            let mut writer = mcap::WriteOptions::new()
-                .emit_summary_offsets(false)
-                .calculate_data_section_crc(false)
-                .calculate_summary_section_crc(false)
-                .calculate_chunk_crcs(false)
-                .library("test-recorder/0.0")
-                .create(&mut output)
-                .expect("writer");
-            let schema_id = writer
-                .add_schema("example", "jsonschema", br#"{"type":"object"}"#)
-                .expect("schema");
-            channel_id = writer
-                .add_channel(schema_id, "/mixed", "json", &BTreeMap::new())
-                .expect("channel");
-            writer
-                .write_to_known_channel(
-                    &MessageHeader {
-                        channel_id,
-                        sequence: 1,
-                        log_time: 10,
-                        publish_time: 10,
-                    },
-                    &[10],
-                )
-                .expect("chunked message");
-            writer.finish().expect("finish");
-        }
-
-        let mut bytes = output.into_inner();
-        let loose_message = message_record(channel_id, 2, 1, &[1]);
-        let data_end_offset = record_offset(&bytes, mcap::records::op::DATA_END);
-        bytes.splice(
-            data_end_offset..data_end_offset,
-            loose_message.iter().copied(),
-        );
-        patch_footer_summary_start(&mut bytes, loose_message.len() as u64);
-        patch_statistics_message_count(&mut bytes, 2);
-        bytes
-    }
-
-    fn message_record(channel_id: u16, sequence: u32, log_time: u64, payload: &[u8]) -> Vec<u8> {
-        let mut body = Vec::new();
-        body.extend_from_slice(&channel_id.to_le_bytes());
-        body.extend_from_slice(&sequence.to_le_bytes());
-        body.extend_from_slice(&log_time.to_le_bytes());
-        body.extend_from_slice(&log_time.to_le_bytes());
-        body.extend_from_slice(payload);
-        wrap_record(mcap::records::op::MESSAGE, &body)
-    }
-
-    fn wrap_record(opcode: u8, body: &[u8]) -> Vec<u8> {
-        let mut record = Vec::with_capacity(9 + body.len());
-        record.push(opcode);
-        record.extend_from_slice(&(body.len() as u64).to_le_bytes());
-        record.extend_from_slice(body);
-        record
-    }
-
-    fn record_offset(bytes: &[u8], target_opcode: u8) -> usize {
-        let mut offset = mcap::MAGIC.len();
-        let records_end = bytes.len() - mcap::MAGIC.len();
-        while offset < records_end {
-            let opcode = bytes[offset];
-            let length =
-                u64::from_le_bytes(bytes[offset + 1..offset + 9].try_into().unwrap()) as usize;
-            if opcode == target_opcode {
-                return offset;
-            }
-            offset += 9 + length;
-        }
-        panic!("record opcode 0x{target_opcode:02x} not found");
-    }
-
-    fn patch_footer_summary_start(bytes: &mut [u8], delta: u64) {
-        let offset = record_offset(bytes, mcap::records::op::FOOTER) + 9;
-        let current = u64::from_le_bytes(bytes[offset..offset + 8].try_into().unwrap());
-        bytes[offset..offset + 8].copy_from_slice(&(current + delta).to_le_bytes());
-    }
-
-    fn patch_statistics_message_count(bytes: &mut [u8], message_count: u64) {
-        let offset = record_offset(bytes, mcap::records::op::STATISTICS) + 9;
-        bytes[offset..offset + 8].copy_from_slice(&message_count.to_le_bytes());
-    }
-
-    fn merge_bytes(
-        inputs: &[(&str, &[u8])],
-        coalesce_channels: CoalesceChannels,
-        allow_duplicate_metadata: bool,
-    ) -> Result<Vec<u8>> {
-        let options = MergeOptions {
-            files: Vec::new(),
-            output_file: None,
-            compression: None,
-            chunk_size: 1024,
-            include_crc: true,
-            chunked: true,
-            allow_duplicate_metadata,
-            coalesce_channels,
-        };
-        let input_refs = inputs
-            .iter()
-            .map(|(name, data)| InputRef { name, data })
-            .collect::<Vec<_>>();
-
-        let mut output = Cursor::new(Vec::<u8>::new());
-        merge_inputs(&input_refs, &mut output, &options, false)?;
-        Ok(output.into_inner())
-    }
+    use super::build_merge_options;
+    use crate::cli::{CoalesceChannels, CompressionFormat, MergeCommand};
 
     #[test]
     fn build_merge_options_maps_cli_fields() {
         let options = build_merge_options(MergeCommand {
             files: vec!["a.mcap".into(), "b.mcap".into()],
-            output_file: Some("out.mcap".into()),
+            output: Some("out.mcap".into()),
+            output_file: None,
             compression: CompressionFormat::Lz4,
             chunk_size: 4096,
             no_crc: true,
@@ -1142,7 +56,7 @@ mod tests {
             options.files,
             vec![PathBuf::from("a.mcap"), PathBuf::from("b.mcap")]
         );
-        assert_eq!(options.output_file, Some(PathBuf::from("out.mcap")));
+        assert_eq!(options.output, Some(PathBuf::from("out.mcap")));
         assert!(matches!(options.compression, Some(mcap::Compression::Lz4)));
         assert_eq!(options.chunk_size, 4096);
         assert!(!options.include_crc);
@@ -1152,727 +66,34 @@ mod tests {
     }
 
     #[test]
-    fn run_rejects_remote_input_without_scan_opt_in() {
-        let err = run(
-            &CommandContext::default(),
-            MergeCommand {
-                files: vec!["http://example.com/a.mcap".into()],
-                output_file: Some("out.mcap".into()),
-                compression: CompressionFormat::Zstd,
-                chunk_size: 1024,
-                no_crc: false,
-                no_chunks: false,
-                allow_duplicate_metadata: false,
-                coalesce_channels: CoalesceChannels::Auto,
-            },
-        )
-        .expect_err("remote merge input should require opt-in");
-
-        assert!(err.to_string().contains("--allow-remote-scan"));
+    fn build_merge_options_prefers_output_over_deprecated_output_file() {
+        let options = build_merge_options(MergeCommand {
+            files: vec!["a.mcap".into()],
+            output: Some("out.mcap".into()),
+            output_file: Some("legacy.mcap".into()),
+            compression: CompressionFormat::Zstd,
+            chunk_size: mcap::WriteOptions::DEFAULT_CHUNK_SIZE,
+            no_crc: false,
+            no_chunks: false,
+            allow_duplicate_metadata: false,
+            coalesce_channels: CoalesceChannels::Auto,
+        });
+        assert_eq!(options.output, Some(PathBuf::from("out.mcap")));
     }
 
     #[test]
-    fn run_rejects_same_input_and_output_without_truncating() {
-        let input = build_mcap("profile", &[], &[], &[], true, true);
-        let dir = tempfile::TempDir::new().expect("temp dir");
-        let path = dir.path().join("same-path.mcap");
-        std::fs::write(&path, &input).expect("write input");
-
-        let err = run(
-            &CommandContext::default(),
-            MergeCommand {
-                files: vec![path.clone()],
-                output_file: Some(path.clone()),
-                compression: CompressionFormat::Zstd,
-                chunk_size: 1024,
-                no_crc: false,
-                no_chunks: false,
-                allow_duplicate_metadata: false,
-                coalesce_channels: CoalesceChannels::Auto,
-            },
-        )
-        .expect_err("same input/output should fail");
-
-        assert!(err.to_string().contains("input and output paths"));
-        assert_eq!(std::fs::read(&path).expect("read input"), input);
-    }
-
-    #[test]
-    fn merge_orders_messages_by_log_time_then_input_index() {
-        let left = build_mcap(
-            "profile",
-            &[
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 5,
-                    payload: vec![1],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![3],
-                },
-            ],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/right".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 3,
-                    payload: vec![2],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/right".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![4],
-                },
-            ],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-        let mut ordered_messages = Vec::<(u64, String, Vec<u8>)>::new();
-        for message in mcap::MessageStream::new(&merged).expect("stream") {
-            let message = message.expect("message");
-            ordered_messages.push((
-                message.log_time,
-                message.channel.topic.clone(),
-                message.data.to_vec(),
-            ));
-        }
-        assert_eq!(
-            ordered_messages,
-            vec![
-                (3, "/right".to_string(), vec![2]),
-                (5, "/left".to_string(), vec![1]),
-                // Tie at log_time=10 resolves by input index: left before right.
-                (10, "/left".to_string(), vec![3]),
-                (10, "/right".to_string(), vec![4]),
-            ]
-        );
-    }
-
-    #[test]
-    fn merge_sorts_unsorted_inputs_and_preserves_tie_order() {
-        let left = build_mcap(
-            "profile",
-            &[
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 20,
-                    payload: vec![5],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 5,
-                    payload: vec![1],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![2],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![3],
-                },
-            ],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/right".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 7,
-                    payload: vec![6],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/right".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![4],
-                },
-            ],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-        let ordered_messages = mcap::MessageStream::new(&merged)
-            .expect("stream")
-            .map(|message| {
-                let message = message.expect("message");
-                (
-                    message.log_time,
-                    message.channel.topic.clone(),
-                    message.data.to_vec(),
-                )
-            })
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            ordered_messages,
-            vec![
-                (5, "/left".to_string(), vec![1]),
-                (7, "/right".to_string(), vec![6]),
-                (10, "/left".to_string(), vec![2]),
-                (10, "/left".to_string(), vec![3]),
-                (10, "/right".to_string(), vec![4]),
-                (20, "/left".to_string(), vec![5]),
-            ]
-        );
-    }
-
-    #[test]
-    fn merge_sorts_non_indexed_unsorted_inputs() {
-        let left = build_non_indexed_mcap(
-            "profile",
-            &[
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 10,
-                    payload: vec![2],
-                },
-                TestMessage {
-                    channel_id: 1,
-                    topic: "/left".to_string(),
-                    metadata: BTreeMap::new(),
-                    log_time: 1,
-                    payload: vec![1],
-                },
-            ],
-        );
-        let right = build_non_indexed_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/right".to_string(),
-                metadata: BTreeMap::new(),
-                log_time: 5,
-                payload: vec![3],
-            }],
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-        let ordered_log_times = mcap::MessageStream::new(&merged)
-            .expect("stream")
-            .map(|message| message.expect("message").log_time)
-            .collect::<Vec<_>>();
-
-        assert_eq!(ordered_log_times, vec![1, 5, 10]);
-    }
-
-    #[test]
-    fn indexed_fast_path_requires_all_messages_in_indexes() {
-        let input = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/demo".to_string(),
-                metadata: BTreeMap::new(),
-                log_time: 1,
-                payload: vec![1],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let mut summary = mcap::Summary::read(&input)
-            .expect("summary")
-            .expect("summary present");
-        assert!(summary_indexes_all_messages(&input, &summary));
-
-        summary.stats.as_mut().expect("stats present").message_count += 1;
-        assert!(!summary_indexes_all_messages(&input, &summary));
-    }
-
-    #[test]
-    fn merge_keeps_loose_messages_when_indexed_summary_is_incomplete() {
-        let input = build_indexed_mcap_with_loose_message();
-
-        let merged = merge_bytes(
-            &[("mixed", input.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-        let messages = mcap::MessageStream::new(&merged)
-            .expect("stream")
-            .map(|message| {
-                let message = message.expect("message");
-                (message.log_time, message.data.to_vec())
-            })
-            .collect::<Vec<_>>();
-
-        assert_eq!(messages, vec![(1, vec![1]), (10, vec![10])]);
-    }
-
-    #[test]
-    fn merge_sets_empty_profile_when_inputs_disagree() {
-        let a = build_mcap("a", &[], &[], &[], true, true);
-        let b = build_mcap("b", &[], &[], &[], true, true);
-
-        let merged = merge_bytes(
-            &[("a", a.as_slice()), ("b", b.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        let header = match mcap::read::LinearReader::new(&merged)
-            .expect("reader")
-            .next()
-            .expect("header")
-            .expect("record")
-        {
-            Record::Header(header) => header,
-            _ => panic!("expected header"),
-        };
-        assert!(header.profile.is_empty());
-    }
-
-    #[test]
-    fn merge_stamps_cli_writer_library() {
-        let a = build_mcap("p", &[], &[], &[], true, true);
-        let b = build_mcap("p", &[], &[], &[], true, true);
-
-        let merged = merge_bytes(
-            &[("a", a.as_slice()), ("b", b.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        // The inputs' `test-recorder/0.0` library is overwritten with the CLI's own identity.
-        let library = crate::parse::read_header(&merged)
-            .expect("read header")
-            .expect("header present")
-            .library;
-        assert_eq!(library, *crate::cli::LIBRARY_IDENTIFIER);
-    }
-
-    #[test]
-    fn merge_rejects_duplicate_metadata_name_by_default() {
-        let first = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("1"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-        let second = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("2"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-
-        let err = merge_bytes(
-            &[("a", first.as_slice()), ("b", second.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect_err("merge should fail");
-        assert!(
-            format!("{err:#}").contains("metadata name 'robot' was previously encountered"),
-            "unexpected error: {err:#}"
-        );
-    }
-
-    #[test]
-    fn merge_allow_duplicate_metadata_deduplicates_identical_records() {
-        let first = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("1"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-        let second = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("1"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("a", first.as_slice()), ("b", second.as_slice())],
-            CoalesceChannels::Auto,
-            true,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.stats.as_ref().expect("stats").metadata_count, 1);
-        assert_eq!(summary.metadata_indexes.len(), 1);
-    }
-
-    #[test]
-    fn merge_allow_duplicate_metadata_keeps_same_name_with_different_content() {
-        let first = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("1"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-        let second = build_mcap(
-            "p",
-            &[],
-            &[mcap::records::Metadata {
-                name: "robot".to_string(),
-                metadata: BTreeMap::from([(String::from("a"), String::from("2"))]),
-            }],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("a", first.as_slice()), ("b", second.as_slice())],
-            CoalesceChannels::Auto,
-            true,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.stats.as_ref().expect("stats").metadata_count, 2);
-        assert_eq!(summary.metadata_indexes.len(), 2);
-    }
-
-    #[test]
-    fn merge_force_coalesces_channels_ignoring_metadata() {
-        let left = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::from([(String::from("host"), String::from("left"))]),
-                log_time: 0,
-                payload: vec![1],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::from([(String::from("host"), String::from("right"))]),
-                log_time: 1,
-                payload: vec![2],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Force,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.channels.len(), 1);
-    }
-
-    #[test]
-    fn merge_auto_keeps_channels_distinct_when_metadata_differs() {
-        let left = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::from([(String::from("host"), String::from("left"))]),
-                log_time: 0,
-                payload: vec![1],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::from([(String::from("host"), String::from("right"))]),
-                log_time: 1,
-                payload: vec![2],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.channels.len(), 2);
-    }
-
-    #[test]
-    fn merge_auto_coalesces_channels_when_metadata_matches() {
-        let metadata = BTreeMap::from([(String::from("host"), String::from("same"))]);
-        let left = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: metadata.clone(),
-                log_time: 0,
-                payload: vec![1],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata,
-                log_time: 1,
-                payload: vec![2],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.channels.len(), 1);
-    }
-
-    #[test]
-    fn merge_none_does_not_coalesce_channels() {
-        let left = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::new(),
-                log_time: 0,
-                payload: vec![1],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-        let right = build_mcap(
-            "profile",
-            &[TestMessage {
-                channel_id: 1,
-                topic: "/topic".to_string(),
-                metadata: BTreeMap::new(),
-                log_time: 1,
-                payload: vec![2],
-            }],
-            &[],
-            &[],
-            true,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[("left", left.as_slice()), ("right", right.as_slice())],
-            CoalesceChannels::None,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.channels.len(), 2);
-    }
-
-    #[test]
-    fn merge_copies_attachments_with_and_without_indexes() {
-        let indexed = build_mcap(
-            "profile",
-            &[],
-            &[],
-            &[mcap::Attachment {
-                log_time: 1,
-                create_time: 1,
-                name: "indexed.bin".to_string(),
-                media_type: "application/octet-stream".to_string(),
-                data: std::borrow::Cow::Borrowed(&[1, 2, 3]),
-            }],
-            true,
-            true,
-        );
-
-        let unindexed = build_mcap(
-            "profile",
-            &[],
-            &[],
-            &[mcap::Attachment {
-                log_time: 2,
-                create_time: 2,
-                name: "unindexed.bin".to_string(),
-                media_type: "application/octet-stream".to_string(),
-                data: std::borrow::Cow::Borrowed(&[4, 5, 6]),
-            }],
-            false,
-            true,
-        );
-
-        let merged = merge_bytes(
-            &[
-                ("indexed", indexed.as_slice()),
-                ("unindexed", unindexed.as_slice()),
-            ],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.stats.as_ref().expect("stats").attachment_count, 2);
-        assert_eq!(summary.attachment_indexes.len(), 2);
-    }
-
-    #[test]
-    fn merge_copies_attachments_without_statistics_record() {
-        let no_stats = build_mcap_with_options(
-            "profile",
-            &[],
-            &[],
-            &[mcap::Attachment {
-                log_time: 3,
-                create_time: 3,
-                name: "nostats.bin".to_string(),
-                media_type: "application/octet-stream".to_string(),
-                data: std::borrow::Cow::Borrowed(&[7, 8, 9]),
-            }],
-            true,
-            true,
-            false,
-        );
-
-        let merged = merge_bytes(
-            &[("no-stats", no_stats.as_slice())],
-            CoalesceChannels::Auto,
-            false,
-        )
-        .expect("merge");
-
-        let summary = mcap::Summary::read(&merged)
-            .expect("summary")
-            .expect("present");
-        assert_eq!(summary.stats.as_ref().expect("stats").attachment_count, 1);
-        assert_eq!(summary.attachment_indexes.len(), 1);
+    fn build_merge_options_falls_back_to_deprecated_output_file() {
+        let options = build_merge_options(MergeCommand {
+            files: vec!["a.mcap".into()],
+            output: None,
+            output_file: Some("legacy.mcap".into()),
+            compression: CompressionFormat::Zstd,
+            chunk_size: mcap::WriteOptions::DEFAULT_CHUNK_SIZE,
+            no_crc: false,
+            no_chunks: false,
+            allow_duplicate_metadata: false,
+            coalesce_channels: CoalesceChannels::Auto,
+        });
+        assert_eq!(options.output, Some(PathBuf::from("legacy.mcap")));
     }
 }
